requirements.txt
altair==4.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.5
astropy==5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==0.7.8
cachetools==5.0.0
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==2.1.0
colorama==0.4.4
cycler==0.11.0
cymem==2.0.6
dask==2022.8.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask==2.1.2
fonttools==4.36.0
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
html5lib==1.1
idna==3.3
imageio==2.21.2
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
it==1.0.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.7.1
locket==1.0.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
multitasking==0.0.11
murmurhash==1.0.8
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.11
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathy==0.6.2
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
plot==0.6.5
pooch==1.6.0
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pt-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.4.0/pt_core_news_md-3.4.0-py3-none-any.whl
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.4.0/pt_core_news_sm-3.4.0-py3-none-any.whl
pure-eval==0.2.2
pyamg==4.2.3
pyarrow==3.0.0
pycparser==2.21
pydantic==1.9.2
pydeck==0.7.1
pyerfa==2.0.0.1
Pygments==2.12.0
pylint==2.13.8
Pympler==1.0.1
pyodbc==4.0.32
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
QtPy==2.2.0
requests==2.27.1
scikit-image==0.19.3
scipy==1.9.0
semver==2.13.0
Send2Trash==1.8.0
SimpleITK==2.2.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-streamlit==1.0.4
srsly==2.4.4
stack-data==0.2.0
streamlit==1.9.0
streamlit-cropper==0.2.0
terminado==0.13.3
thinc==8.1.0
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
wasabi==0.10.1
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
yfinance==0.1.74
zipp==3.8.0